Change default action to show
Previous default action was to increment MAC by one.  This default was not
documented and is "dangerous" (changes MAC, change is rather subtle and easy to
miss).

This changes the default action to --show, which has already been the default
in Debian packages for a while, see:

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=621494

Upstream bug report:
https://github.com/alobbs/macchanger/issues/4