Added option to force a different mac address
Occasionally, a random address can be the same as the previous one.
That patch adds a --different option that will try again if the we are unlucky
enough to get a random one that is the same one.
Some distributions such as tails are paranoid enough not to trust the exit code
and will take drastic measures if macchange doesn't actually change the mac
address. That option should help them.